Fix schema not added when subscribe an empty topic without schema (#9853)
Fixes #7728

When a consumer with a schema subscribes an empty topic without schema, the current check uses `isActive`, which only checks if the topic can be deleted. However, it should check if there's any connected producer or consumer of this topic. For current implementation, even if a topic has no active producers of consumers, the topic's subscription list may be not empty and `isActive` will return true. Then the schema of consumer won't be attached to the topic and it will throw an `IncompatibleSchemaException`.

- Check if the topic has active producers or consumers instead of whether it can be deleted.
- Add related tests.

(cherry picked from commit 7f92adee9d9d60f92d14896b52c885eecfde4d77)