[Build] Skip python 3.9 for master branch ---
Master Issue: #9888

*Motivation*

We haven't publish the image `apachepulsar/pulsar-build:manylinux-3.9`.
This will break our currently master build python wheel file.
I was trying to build out the pulsar build image with python 3.9 but
that can not done very quickly. So we can skip the python3.9 build and
after we build out the base image then we can add it back.