[Issue 9602] Add schema type validation (#9797)
Fixes #9602

The schema should not be able to modify the type after it is first created. Unless he is Always compatible.

When uploading a schema, it is verified that the type of the newly uploaded schema is consistent with the schema in the metadata.

When checkCompatible, change the Primitive type from checking only to checking all.

* When a producer is created, the schema is defined, and when a producer is created again, the schema type is not always the same, otherwise an exception will be thrown.

* When a producer is created, the schema is defined, and when a consumer is created again, the schema type is not always the same, otherwise an exception will be thrown.

* When a consumer is created, the schema is defined, and when a producer is created again, the schema type is not always the same, otherwise an exception will be thrown.

* When a consumer is created, the schema is defined, and when a consumer is created again, the schema type is not always the same, otherwise an exception will be thrown.

* When compatibility is set to always compatible. No exceptions are thrown.

*If `yes` was chosen, please highlight the changes*
  - The schema: (yes)

(cherry picked from commit b45added9d27adf5218ca005a82f4faa18b7ec44)