Fix the useless retry when the maximum number of subscriptions is reached (#9991)
When the maximum number of subscriptions is reached, due to the long default value of operationTimeout, the client will keep retrying and get stuck for a long time.
If a NotAllowedException is returned, the client will not continue to retry and will fail directly.

Can the test fail in a short time

(cherry picked from commit 31c5fb52bc841e40d678bd3787269160b7d14c12)