[Documentation][Geo Replication] Document clusters create step in geo replication (#9857)
### Motivation

The geo replication documentation should reference the way to configure connections among clusters, as they are essential to configuring geo replication.

### Modifications

Added documentation to the primary site2 documentation.

### Question

I think it'd be valuable to update the documentation for earlier versions of pulsar, but I haven't looked into it yet. If these changes are accepted, I'm happy to discuss which versions to update.

### Acknowledgements

I want to acknowledge that the content for this documentation update is motivated by @pckeyan's blog: https://medium.com/@pckeyan/apache-pulsar-geo-replication-ad4f0ca3224b. The change is not a straight copy from the blog, though.