pulsar-perf: add ability to create partitioned topics (#9859)
### Motivation
When you are using pulsar-perf you have to pre-create the topic if you want it to be partitioned.
This is very awkward and error prone.

### Modifications

This patch adds the ability to set a number of partitions.
Before starting the main procedure we are going to create the topics as partitioned topics.
I needed to add support to pass explicitly the http service url, because by default the serverURL is the binary protocol URL and it is not suitable for PulsarAdmin. The default value is http://localhost:8080.

I added support for not calling System.exit and execute a custom callback, otherwise it is not possible to execute reliably the tests without crashing the JVM.

### Verifying this change

This change added tests and can be verified bu testing manually the new option.