[Issue 10058]:apply-config-from-env.py to commented default values (#10060)
Add handling of commented name=value so that apply-config-from-env.py may un-comment the values if needed.

It is common practice to leave default values commented out in the configuration files. By not handling the commented case, the end customer may need to jump through hoops(such as adding un-commented values via script)  to make changes to these values in a production environment. This enhancement will ease the process by allowing the script to catalog the commented as well as un-commented values, and make replacements to both as needed, leading to a smoother customer experience when a support situation calls for adjusting the value to a commented variable. 

If a value is not supplied to  a commented variable, then the line remains commented out, and the default value should be assumed by the application. 

Any lines that do not contain a = or result in a parsing error should be ignored and the original line left in the configuration file.