Update chuckedMessageRate to chunkedMessageRate without breaking public api (#10223)
Fixes #10046  

### Motivation

As described in #10046, there is a typo where `chuck` is used instead of `chunk` in several locations, including in the public API. This PR fixes that typo without introducing any breaking changes. Instead, it adds some new methods while deprecating ones that contain typos.

### Modifications

1. Update several internal pulsar references. Given that these are private, there is no need to worry about breaking changes.
2. Update `ConsumerBuilder` interface by first deprecating the `maxPendingChuckedMessage` method and adding the `maxPendingChunkedMessage` method. We'll need to choose when to actually remove the deprecated method.
3. Update `SubscriptionStats` to return both `chuckedMessageRate` and `chunkedMessageRate`. Given that this POJO is sent as JSON, the only way to give users the chance to transition is to deliver both fields.
4. Update `ConsumerStats` to return both `chuckedMessageRate` and `chunkedMessageRate`. Given that this POJO is sent as JSON, the only way to give users the chance to transition is to deliver both fields.
5. Update `ConsumerConfigurationData` by changing the private variable from `maxPendingChuckedMessage` to `maxPendingChunkedMessage`. Then, add the previously generated getter and setter for `maxPendingChuckedMessage` that actually update `maxPendingChunkedMessage`.

### Verifying this change

I updated several tests but didn't add any. I'm not sure what the protocol is for these types of changes. The public API is not changing, and in the case of deprecated methods, any affected tests are updated to use the new methods.

### Does this pull request potentially affect one of the following parts:

*If `yes` was chosen, please highlight the changes*

  - Dependencies (does it add or upgrade a dependency): no
  - The public API: yes 
  - The schema: no
  - The default values of configurations: no
  - The wire protocol: no
  - The rest endpoints: no
  - The admin cli options: no
  - Anything that affects deployment: no

The changes to the public API are only additive.

### Documentation

There is updated documentation associated with this change. If we want to merge this change to the 2.7 branch, we'll want to update that documentation as well.