Improve localrun performance by using JVM options more suited for it (#10273)
### Motivation

Currently, the JVM options used for localrun are the same as that used for the pulsar-admin tool.  The JVM options used are suited for CLI tools that need a short startup time.  Sinkce localrun is a long running process, we can achieve better performance especially in GC if we use JVM options more suited for this kind of workload.  I am seeing around 20% better performance with these changes in JVM options.