Pause the replicated subscriptions snapshots when there is no real traffic (#10292)
### Motivation

Fixes #6437

This is a different approach to fix the issue, alternative to #7299. 

Instead of automatically acking the marker messages when the consumers are not connected, we should instead stop taking the snapshot when there's no traffic. There's no point in continuing storing these snapshots all the time if there's no traffic.

### Modifications

 1. When we detect that there has been no new messages since the last snapshot was take, skip the new snapshot.
 2. When there are no local producers, skip 1 snapshot creation. This is done in order to create a quiet period and give more time to not have any remote markers since our last snapshot was created.