[broker] Limit the number of producers/consumers that can connect per topic for each IP address (#10188)
### Motivation

Occasionally, a client application developer makes an implementation mistake and increases the number of producers or consumers infinitely. If this happens, there is concern that all file descriptors on the broker server will be exhausted.

Brokers already have settings such as `maxProducersPerTopic` and `maxConsumersPerTopic`, which allow us to specify a maximum number of clients that can connect per topic. However, if one topic is used by multiple tenants, there is an issue that if one tenant accidentally increases the number of connected clients too much, other tenants will not be able to add new clients.

### Modifications

Enable users to limit the number of producers or consumers which can connect to each topic **with the same IP address**.