Fix ConcurrentModificationException when attempting to update local broker data (#10347)
Fixes #10229

### Motivation
`updateLocalBrokerData` will be called concurrently in the REST API
`writeBrokerDataOnZooKeeper` will be called concurrently in REST API and Scheduled Task

It is not enough to ensure the thread safety of every API in LocalBrokerData.
It is necessary to ensure the thread safety of multiple APIs in  LocalBrokerData at the same time. 
So I added the lock to the two methods above.