[Functions] Adding “numListenerThreads” configuration parameter in WorkerConfig . #8575
The reason for this problem is that there is only one PulsarClient in a functionWorker instance,
and the number of externalExecutorProvider thread pools in PulsarClient is 1,
so all consumers share the same listener executor.
Therefore, when one or more tasks stoped, the listener executor will be blocked, and finally other tasks will stop consuming data.