[Branch-2.7] Fix deadlock on Monitoring thread blocked by LeaderService.isLeader() (#10512)
---

Fixes #10235

According to #10235, when `LeaderService` is changing leadership status
(like losing leadership, or becoming a leader), the `LeaderService` will
be locked with `synchronized` block. Which will block other threads if
calling `LeaderService.isLeader()`. This PR changes `ClusterServiceCoordinator`
and `WorkerStatsManager` to check if is leader from `MembershipManager`,
which will not block other threads if `LeaderService` is at `synchronized` block.

Also, this PR will not resolve the root cause of #10235, since there is
lack of context about blocked reader for the `FunctionAssignmentTailer`.

- [ ] Make sure that the change passes the CI checks.

--- 
Original PR is https://github.com/apache/pulsar/pull/10502