Fix the inconsistency of AdvertisedAddress (#10312)
### Motivation
There is an `advertisedAddress` in PusarService and an `advertisedAddress` in conf. When `advertisedListener` is set, the values of the two `advertisedAddress` will be inconsistent. 
In the current code, `PusarService.getAdvertisedAddress` is used in some places, and `conf.getAdvertisedAddress` is used in other places.

### Modifications
Put this logic into the config to keep the values in all places consistent.