Fix the inconsistency of AdvertisedAddress (#10312)
There is an `advertisedAddress` in PusarService and an `advertisedAddress` in conf. When `advertisedListener` is set, the values of the two `advertisedAddress` will be inconsistent.
In the current code, `PusarService.getAdvertisedAddress` is used in some places, and `conf.getAdvertisedAddress` is used in other places.

Put this logic into the config to keep the values in all places consistent.

(cherry picked from commit d39e5e41ab37423c7d7d9e71d5c5fd7fc2c03d5c)