[Functions] Process async results in the same Java runnable thread
*Motivation*

After introducing the support for async functions, the java function processing semantic is not enforced.
For example, if it fails to write a sink, it doesn't fail the java instance or fail the message. Hence it keeps
receiving messages but never ack or nack.

*Modification*

Change the way how aysnc function requests are processed to fix the issues we have seen in Kinesis connector.