[Functions] Process async results in the same Java runnable thread (#10618)
* [Functions] Process async results in the same Java runnable thread

*Motivation*

After introducing the support for async functions, the java function processing semantic is not enforced.
For example, if it fails to write a sink, it doesn't fail the java instance or fail the message. Hence it keeps
receiving messages but never ack or nack.

*Modification*

Change the way how aysnc function requests are processed to fix the issues we have seen in Kinesis connector.

* Process the async results in the different thread

(cherry picked from commit 1cbd5b4dbb2c7611650065bb7be0a3541cf48b50)