Release OpAddEntry.data when entry is copied and discarded (#10773)
Fixes #10738

### Motivation

There's a ByteBuf leak that happens in ledger rollover. The repro case in #10738 reproduces some Netty ByteBuf leaks that are detected by the Netty Leak detector.

### Modifications

Release `OpAddEntry.data` when the original entry is copied and discarded

### Additional Context

This PR was split out of PR #10755 which was closed without merging.