Revert "Enable pulsar function to send message to external pulsar clu… (#10843)
This reverts commit 66231e313502bfe74fd28976560d4391c5cdefa5.

### Motivation

Reverting since the external cluster configuration was not not needed as there are better ways to interact with external pulsar clusters.  A user can simply instantiate a Pulsar client, producer, or consumer in a function to interact with another pulsar cluster.  This feature is also a security risk as it does not security handle authentication data and it will be written as plain text in the internal function's metadata topic.