[Issue 10860][pulsar-metadata] Ensure metadata cache consistency across brokers (#10862)
* Added metadata cache test to simulate multi broker cache

* fix create and delete ops on cache

1. During create we should add a watch for the path in zookeeper. Without this
we will not be notified if the znode is changed on another brokers

2. similarly when deleting, the cache should be invalidated. But we shouldn't add an
entry to the cache. This could get added again on some other broker. In that
case we need to go a fetch the entry from the zookeeper. Adding an empty
entry to the cache prevents that.

* Address review comments

Also add a small delay in test to allow notifications to propagate to other
caches. Without this the tests are occasionally failing

Co-authored-by: Surinder Singh <surinders@splunk.com>