Fix compaction not working for system topic (#10941)
If a topic only have non-durable subscriptions but not durable subscriptions,
and the non-durable subscription reach the end of the topic, we will get 0 estimated backlog size
So that the compaction will never been triggered.

The expected behavior is if we have no durable subscriptions, we should use the total size for triggering
the compaction.

(cherry picked from commit 797cb128045c1f2bf304c7de7454f1f12f8312e7)