Fix NonRecoverableLedgerException when get last message ID by Reader
If a topic only have non-durable subscriptions or mark delete position of all the durable subscriptions
are reached the LAC, all the ledgers except the current ledger will be deleted.
Since the current ledger may not have any data, occurs NonRecoverableLedgerException in the broker side.

In this case, the we should return the message ID (-1, -1).