Fix potential data lost on the system topic when topic compaction have not triggered yet.
To trigger the system topic compaction to avoid lost any data since we are using reader for reading
data from the __change_events topic, if no durable subscription on the topic, the data might be lost.
Since we are using the topic compaction on the __change_events topic to reduce the topic policy
cache recovery time, so we can leverage the topic compaction cursor for retaining the data.

We can trigger the compaction directly here since for the first time we create the system topic,
The system topic does not have any data, the operation just pre-create the subscription for the
topic compaction.