Update SQL config to use composed contaiiners
When we use docker-compose to create multiple service containers we need
to override the configuration of each of these containers to use the
correct URLs for the other containers running the other Pulsar services.
The sql1 container must be configured to use the correct broker, webservice
and zookeeper containers.

Added some documentation to the apply-config-from-env-with-prefix.py
script.