Fix potential data lost on the system topic when topic compaction have not triggered yet (#11003)
To pre-create the subscription for the compactor to avoid lost any data since we are using reader
for reading data from the __change_events topic, if no durable subscription on the topic,
the data might be lost. Since we are using the topic compaction on the __change_events topic
to reduce the topic policy cache recovery time,
so we can leverage the topic compaction cursor for retaining the data.