[broker] Add REST API to enable or disable replicated subscriptions (#10790)
### Motivation

Currently, once a replicated subscription is created on a topic, there is no way for users to disable it. I think it is useful to have a REST API that allows users to enable or disable replicated subscriptions.

### Modifications

Added the following REST API endpoints:
```
/admin/persistent/{tenant}/{cluster}/{namespace}/{topic}/subscription/{subName}/replicatedSubscriptionStatus
/admin/v2/persistent/{tenant}/{namespace}/{topic}/subscription/{subName}/replicatedSubscriptionStatus
```
We can enable/disable a replicated subscription by posting true or false to these endpoints.