[ Issues 10788 ] Fixed  flaky test AdminApiTest.
See this issue: https://github.com/apache/pulsar/issues/10788

Due to `NamespaceService` split and `updateNamespaceBundles` may fail because of concurrent write to zookeeper,
And exceeded the maximum number of retries.
Source code is here:
> NamespaceService 838:856

Since this error is less likely to occur, we can try to retry the test in case of an error to solve the problem.