Fix subscription permission not working in reset cursor (#11132)
### Motivation

Some `internalResetCursorXX` methods do not pass in the `subscriptionName` parameter when verifying permissions, which causes the `subscription` check to be skipped during the permission check of `AuthorizationProvider#canConsumeAsync` and leads an error validation result. This PR will fix this problem.

### Modifications

Refine the parameters of `validateTopicOperation` and supplement a relative test case.