Fix reading compacted data with the earliest start position by using Reader API
Currently, if the data has been compacted and the reader try to read from the start of the topic,
the read not able to read any data due to the start position will be change when recovery the NonDurableCursor.
The start position will changed to the first position of the ledgers, but there are some more earlier data in the
compacted ledger. So the reader will lost the compacted data.

The fix is adding a force start position option for creating a NonDurableCursor.

1. If there is compacted context, use the first entry of the compacted ledger as the start position
2. Force seek to a position without compare with the mark deletion position / LAC when reading data from the compacted ledger.
3. Don't rewind the non-durable cursor during the message reading