[Branch-2.7] On multi-topic consumer, we shouldn't keep checking the partitioned metadata  (#11168)
Pick #10708 to branch 2.7.

### Motivation

When using a multi-topic consumer, either through the regex or by passing a list of topics, we are ending up checking the partitioned topic metadata every 1 min, even if the topics are not partitioned. 

We don't need to re-check that information if the topics are not partitioned. Also, this can be causing a lot of read traffic on ZK since we keep checking that "partitioned-metadata" that these topics don't have and, in 2.7, this information is not cached in brokers when the metadata is not found. 

### Modifications

When adding non-partitioned topics to the multi-topic consumer, use 0 for partitions when the topic is not partitioned in order to skip the periodic re-check of the partitions count.