Disable replicate system topic across clusters.
Currently, we are not able to use the topic policy for a geo-replicated schema.
The root cause is we are not support replicate schema to the remote cluster but the
topic policy using the AVRO schema.

For the schema replication we already have PIP 88 in 2.9.0
For the topic policy replication, we are working on a proposal for now, will available in 2.9.0

But for 2.7.x and 2.8.x, we need this fix to make sure the topic policy can work on a geo replication cluster.