Invalidate the read handle after all cursors consumed.
Currently, the read ReadHandle only invalidate when removing the ledger from the ManagedLedger.
If the ManagedLedger have many ledgers(the topic might retain infinite data), we will get oom
on the direct memory since we are using 1MB read cache by default for the offloaded data ReadHandle.

If all the cursors are consumed the data, the ReadHandle can be closed safety. And if a cursor reset
to an earlier position to consume the historical data, the ReadHandle will be reopen again.